+++ ShoppingCart/ShoppingCart.cs
@@ -1,11 +1,7 @@
 ﻿namespace StardewMods.ShoppingCart;
 
-using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
-using System.Reflection.Emit;
-using HarmonyLib;
 using StardewModdingAPI.Events;
 using StardewModdingAPI.Utilities;
 using StardewMods.Common.Helpers;
@@ -24,6 +20,7 @@ public class ShoppingCart : Mod
     private readonly PerScreen<bool> _makePurchase = new();
 
     private IReflectedField<string?>? _boldTitleText;
+    private ModConfig? _config;
     private IReflectedMethod? _getHoveredItemExtraItemAmount;
     private IReflectedMethod? _getHoveredItemExtraItemIndex;
 
@@ -57,6 +54,21 @@ public class ShoppingCart : Mod
 
     private static string? HoverText => ShoppingCart.Instance!._hoverText?.GetValue();
 
+    private ModConfig Config => this._config ??= CommonHelpers.GetConfig<ModConfig>(this.Helper);
+
+    /// <summary>
+    ///     Check if current menu supports ShoppingCart.
+    /// </summary>
+    /// <param name="menu">The menu to check.</param>
+    /// <returns>Returns true if menu is supported.</returns>
+    public static bool IsSupported(IClickableMenu? menu)
+    {
+        return menu is ShopMenu { currency: 0, storeContext: not ("Dresser" or "FishTank") } shopMenu
+            && shopMenu.forSale.OfType<Item>().Any()
+            && !(shopMenu.portraitPerson?.Equals(Game1.getCharacterFromName("Clint")) == true
+              && shopMenu.forSale.Any(forSale => forSale is Axe or WateringCan or Pickaxe or Hoe or GenericTool));
+    }
+
     /// <inheritdoc />
     public override void Entry(IModHelper helper)
     {
@@ -64,44 +76,15 @@ public class ShoppingCart : Mod
         Log.Monitor = this.Monitor;
         I18n.Init(this.Helper.Translation);
         Integrations.Init(this.Helper);
+        Patches.Init(this.Helper, this.ModManifest, this.Config);
 
         // Events
         this.Helper.Events.Display.MenuChanged += this.OnMenuChanged;
         this.Helper.Events.Display.RenderedActiveMenu += ShoppingCart.OnRenderedActiveMenu;
+        this.Helper.Events.GameLoop.GameLaunched += this.OnGameLaunched;
         this.Helper.Events.Input.ButtonPressed += this.OnButtonPressed;
         this.Helper.Events.Input.CursorMoved += ShoppingCart.OnCursorMoved;
         this.Helper.Events.Input.MouseWheelScrolled += ShoppingCart.OnMouseWheelScrolled;
-
-        // Patches
-        var harmony = new Harmony(this.ModManifest.UniqueID);
-        harmony.Patch(
-            AccessTools.Method(typeof(InventoryMenu), nameof(InventoryMenu.leftClick)),
-            postfix: new(typeof(ShoppingCart), nameof(ShoppingCart.InventoryMenu_leftClick_postfix)));
-        harmony.Patch(
-            AccessTools.Method(typeof(InventoryMenu), nameof(InventoryMenu.rightClick)),
-            postfix: new(typeof(ShoppingCart), nameof(ShoppingCart.InventoryMenu_rightClick_postfix)));
-        harmony.Patch(
-            AccessTools.Constructor(
-                typeof(ShopMenu),
-                new[]
-                {
-                    typeof(List<ISalable>),
-                    typeof(int),
-                    typeof(string),
-                    typeof(Func<ISalable, Farmer, int, bool>),
-                    typeof(Func<ISalable, bool>),
-                    typeof(string),
-                }),
-            transpiler: new(typeof(ShoppingCart), nameof(ShoppingCart.ShopMenu_constructor_transpiler)));
-        harmony.Patch(
-            AccessTools.Method(typeof(ShopMenu), nameof(ShopMenu.receiveScrollWheelAction)),
-            new(typeof(ShoppingCart), nameof(ShoppingCart.ShopMenu_receiveScrollWheelAction_prefix)));
-        harmony.Patch(
-            AccessTools.Method(typeof(ShopMenu), "tryToPurchaseItem"),
-            new(typeof(ShoppingCart), nameof(ShoppingCart.ShopMenu_tryToPurchaseItem_prefix)));
-        harmony.Patch(
-            AccessTools.Method(typeof(ShopMenu), nameof(ShopMenu.updatePosition)),
-            postfix: new(typeof(ShoppingCart), nameof(ShoppingCart.ShopMenu_updatePosition_postfix)));
     }
 
     /// <inheritdoc />
@@ -118,137 +101,6 @@ public class ShoppingCart : Mod
     private static int GetHoveredItemExtraItemIndex()
     {
         return ShoppingCart.Instance!._getHoveredItemExtraItemIndex?.Invoke<int>() ?? -1;
-    }
-
-    [SuppressMessage("ReSharper", "InconsistentNaming", Justification = "Harmony")]
-    [SuppressMessage("StyleCop", "SA1313", Justification = "Harmony")]
-    private static void InventoryMenu_leftClick_postfix(
-        InventoryMenu __instance,
-        ref Item? __result,
-        int x,
-        int y,
-        Item? toPlace)
-    {
-        if (ShoppingCart.CurrentShop is null || toPlace is not null || __result is null)
-        {
-            return;
-        }
-
-        var component = __instance.inventory.Single(cc => cc.containsPoint(x, y));
-        var slotNumber = int.Parse(component.name);
-
-        if (Integrations.StackQuality.IsLoaded && __result is SObject obj)
-        {
-            var stacks = Integrations.StackQuality.API.GetStacks(obj);
-            for (var i = 0; i < 4; ++i)
-            {
-                Item? split = null;
-                var take = new int[4];
-                take[i] = stacks[i];
-                if (stacks[i] <= 0 || !Integrations.StackQuality.API.SplitStacks(obj, ref split, take))
-                {
-                    continue;
-                }
-
-                ShoppingCart.CurrentShop.AddToCart(split);
-            }
-
-            // Return item to inventory
-            Integrations.StackQuality.API.UpdateQuality(obj, stacks);
-            if (__instance.actualInventory[slotNumber] is SObject otherObj)
-            {
-                otherObj.addToStack(obj);
-                __result = null;
-                return;
-            }
-
-            __instance.actualInventory[slotNumber] = obj;
-            __result = null;
-            return;
-        }
-
-        if (!ShoppingCart.CurrentShop.AddToCart(__result))
-        {
-            return;
-        }
-
-        // Return item to inventory
-        __instance.actualInventory[slotNumber] = __result;
-        __result = null;
-    }
-
-    [SuppressMessage("ReSharper", "InconsistentNaming", Justification = "Harmony")]
-    [SuppressMessage("StyleCop", "SA1313", Justification = "Harmony")]
-    private static void InventoryMenu_rightClick_postfix(
-        InventoryMenu __instance,
-        ref Item? __result,
-        int x,
-        int y,
-        Item? toAddTo)
-    {
-        if (ShoppingCart.CurrentShop is null || toAddTo is not null || __result is null)
-        {
-            return;
-        }
-
-        var component = __instance.inventory.Single(cc => cc.containsPoint(x, y));
-        var slotNumber = int.Parse(component.name);
-
-        if (Integrations.StackQuality.IsLoaded && __result is SObject obj)
-        {
-            var stacks = Integrations.StackQuality.API.GetStacks(obj);
-            for (var i = 0; i < 4; ++i)
-            {
-                Item? split = null;
-                var take = new int[4];
-                take[i] = stacks[i];
-                if (stacks[i] <= 0 || !Integrations.StackQuality.API.SplitStacks(obj, ref split, take))
-                {
-                    continue;
-                }
-
-                ShoppingCart.CurrentShop.AddToCart(split);
-            }
-
-            // Return item to inventory
-            Integrations.StackQuality.API.UpdateQuality(obj, stacks);
-            if (__instance.actualInventory[slotNumber] is SObject otherObj)
-            {
-                otherObj.addToStack(obj);
-                __result = null;
-                return;
-            }
-
-            __instance.actualInventory[slotNumber] = obj;
-            __result = null;
-            return;
-        }
-
-        if (!ShoppingCart.CurrentShop.AddToCart(__result))
-        {
-            return;
-        }
-
-        // Return item to inventory
-        var slot = __instance.actualInventory.ElementAtOrDefault(slotNumber);
-        if (slot is not null)
-        {
-            slot.Stack += __result.Stack;
-        }
-        else
-        {
-            __instance.actualInventory[slotNumber] = __result;
-        }
-
-        __result = null;
-    }
-
-    private static bool IsSupported(IClickableMenu? menu)
-    {
-        return menu is ShopMenu { currency: 0, storeContext: not ("Dresser" or "FishTank") } shopMenu
-            && shopMenu.forSale.OfType<Item>().Any()
-            && !(shopMenu.portraitPerson?.Equals(Game1.getCharacterFromName("Clint")) == true
-              && shopMenu.forSale.Any(forSale => forSale is Axe or WateringCan or Pickaxe or Hoe or GenericTool));
     }
 
     private static void OnCursorMoved(object? sender, CursorMovedEventArgs e)
@@ -325,79 +177,6 @@ public class ShoppingCart : Mod
         ShoppingCart.CurrentMenu.drawMouse(e.SpriteBatch);
     }
 
-    private static IEnumerable<CodeInstruction> ShopMenu_constructor_transpiler(
-        IEnumerable<CodeInstruction> instructions)
-    {
-        foreach (var instruction in instructions)
-        {
-            if (instruction.Calls(AccessTools.Method(typeof(ShopMenu), nameof(ShopMenu.updatePosition))))
-            {
-                yield return new(OpCodes.Ldarg_3);
-                yield return new(OpCodes.Ldarg_S, (short)6);
-                yield return CodeInstruction.Call(typeof(ShoppingCart), nameof(ShoppingCart.ShopMenu_updatePosition));
-            }
-            else
-            {
-                yield return instruction;
-            }
-        }
-    }
-
-    private static bool ShopMenu_receiveScrollWheelAction_prefix()
-    {
-        var (x, y) = Game1.getMousePosition(true);
-        return ShoppingCart.CurrentShop?.Bounds.Contains(x, y) != true;
-    }
-
-    [SuppressMessage("ReSharper", "InconsistentNaming", Justification = "Harmony")]
-    [SuppressMessage("StyleCop", "SA1313", Justification = "Harmony")]
-    private static bool ShopMenu_tryToPurchaseItem_prefix(
-        ShopMenu __instance,
-        ref bool __result,
-        ISalable item,
-        ISalable? held_item,
-        int numberToBuy)
-    {
-        if (ShoppingCart.CurrentShop is null
-         || ShoppingCart.MakePurchase
-         || held_item is not null
-         || __instance.readOnly
-         || !ShoppingCart.CurrentShop.AddToCart(item, numberToBuy))
-        {
-            return true;
-        }
-
-        __result = false;
-        return false;
-    }
-
-    private static void ShopMenu_updatePosition(ShopMenu shopMenu, string who, string context)
-    {
-        shopMenu.updatePosition();
-
-        if (who is not "ClintUpgrade" && context is not ("Dresser" or "FishTank"))
-        {
-            return;
-        }
-
-        shopMenu.xPositionOnScreen += VirtualShop.MenuWidth / 2;
-        shopMenu.upperRightCloseButton.bounds.X -= VirtualShop.MenuWidth / 2 + Game1.tileSize;
-    }
-
-    [SuppressMessage("ReSharper", "InconsistentNaming", Justification = "Harmony")]
-    [SuppressMessage("ReSharper", "SuggestBaseTypeForParameter", Justification = "Harmony")]
-    [SuppressMessage("StyleCop", "SA1313", Justification = "Harmony")]
-    private static void ShopMenu_updatePosition_postfix(ShopMenu __instance)
-    {
-        if (!ShoppingCart.IsSupported(__instance))
-        {
-            return;
-        }
-
-        __instance.xPositionOnScreen -= VirtualShop.MenuWidth / 2;
-        __instance.upperRightCloseButton.bounds.X += VirtualShop.MenuWidth / 2 + Game1.tileSize;
-    }
-
     private void OnButtonPressed(object? sender, ButtonPressedEventArgs e)
     {
         if (ShoppingCart.CurrentShop is null
@@ -418,6 +197,26 @@ public class ShoppingCart : Mod
         }
 
         this.Helper.Input.Suppress(e.Button);
+    }
+
+    private void OnGameLaunched(object? sender, GameLaunchedEventArgs e)
+    {
+        if (!Integrations.GMCM.IsLoaded)
+        {
+            return;
+        }
+
+        Integrations.GMCM.Register(
+            this.ModManifest,
+            () => this._config = new(),
+            () => this.Helper.WriteConfig(this.Config));
+
+        Integrations.GMCM.API.AddNumberOption(
+            this.ModManifest,
+            () => this.Config.ShiftClickQuantity,
+            value => this.Config.ShiftClickQuantity = value,
+            I18n.Config_ShiftClickQuantity_Name,
+            I18n.Config_ShiftClickQuantity_Tooltip);
     }
 
     private void OnMenuChanged(object? sender, MenuChangedEventArgs e)